Restructure users
